# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GO API ESHOP Docs.",
        "title": "GO API ESHOP",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "ESHOP API Support",
            "email": "contact@owonafx.com"
        },
        "license": {
            "name": "api-eshop 2.0"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/auth/products": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "CreateProduct create a new Product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "create a new Product",
                "operationId": "create-product",
                "parameters": [
                    {
                        "description": "ProductRequest",
                        "name": "ProductRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createProductRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.productResponse"
                        }
                    },
                    "400": {
                        "description": "Status BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Status Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Status Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Status UnprocessableEntity",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Status BadGateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/products/all/{limit}/{offset}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Display all Products. Auth is required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Display all Products",
                "operationId": "all-products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit number of products returned (default is 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset/Skip number of products (default is 0)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.productListResponse"
                        }
                    },
                    "400": {
                        "description": "Status BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Status Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Status BadGateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/products/top/{limit}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Display top Products. Auth is required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Display top Products",
                "operationId": "top-products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit number of products returned (default is 20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.productListResponse"
                        }
                    },
                    "400": {
                        "description": "Status BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Status Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Status BadGateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/products/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a product. Auth required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get a product",
                "operationId": "get-product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of the product to get",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.productResponse"
                        }
                    },
                    "400": {
                        "description": "Status BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Status Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a Product. Auth is required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Update a new Product",
                "operationId": "update-product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of the product to update",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Product to update",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateProductRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.productResponse"
                        }
                    },
                    "400": {
                        "description": "Status BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Status Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Status Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Status Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Status UnprocessableEntity",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Status BadGateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete a Product. Auth is required",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Delete a Product",
                "operationId": "delete-product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of the product to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.productResponse"
                        }
                    },
                    "400": {
                        "description": "Status BadRequest",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Status Forbidden",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Status Not Found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "409": {
                        "description": "Status Conflict",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "422": {
                        "description": "Status UnprocessableEntity",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "502": {
                        "description": "Status BadGateway",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.createProductRequest": {
            "type": "object",
            "required": [
                "brand",
                "category",
                "description",
                "name",
                "price"
            ],
            "properties": {
                "brand": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "countInStock": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "description": "UserID      uint   `json:\"user_id\" validate:\"required\"`",
                    "type": "string"
                },
                "numReviews": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "rating": {
                    "description": "Reviews      []Review `json:\"reviewList\"`",
                    "type": "number"
                }
            }
        },
...
        "handlers.productListResponse": {
            "type": "object",
            "properties": {
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.productResponse"
                    }
                },
                "productsCount": {
                    "type": "integer"
                }
            }
        },
        "handlers.productResponse": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "countInStock": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "description": "UserID      uint   `json:\"user_id\" validate:\"required\"`",
                    "type": "string"
                },
                "numReviews": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "rating": {
                    "description": "Reviews      []Review",
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}